Extract CategoryIcons to separate component
